refactor: move renderCard to separate file
